requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
google-generativeai>=0.3.0
pytz>=2023.3
sendgrid>=6.10.0
boto3>=1.34.0
lxml>=4.9.0